use uint32_t instead of size_t

+++ include/distingnt/api.h
@@ -518,7 +518,7 @@ void		NT_sendMidi3ByteMessage( uint32_t destination, uint8_t b0, uint8_t b1, uin
 // Sends a System Exclusive message.
 // NT_sendMidiSysEx() may be called repeatedly with partial messages, with 'end' set to false.
 // Set 'end' to true for the final call. You do not need to include the 0xF7 byte to end the SysEx message.
-void		NT_sendMidiSysEx( uint32_t destination, const uint8_t* data, size_t count, bool end );
+void		NT_sendMidiSysEx( uint32_t destination, const uint8_t* data, uint32_t count, bool end );
 
 }
